refactor: move type-name canonicalization to name_classification (C10 stage-2)
`model.py` still mixed the pure type-name string heuristics
(`canonicalize_type_name`, `cv_qualifiers_only_differ`, and the private
`_strip_cv_qualifiers` / `_has_top_level_ptr_or_ref` helpers + their regexes)
with the core data model. These are the same family as the symbol-/type-name
predicates already relocated to the dependency-free `name_classification` leaf in
C10 stage-1.

Move them to `name_classification` (changing the `_re` alias to its plain `re`
import) and re-export `canonicalize_type_name` / `cv_qualifiers_only_differ`
from `model` via the established redundant-alias idiom, so the historical
`from .model import …` path used by the diff modules keeps working unchanged.
`model.py`'s now-unused `import re as _re` is dropped.

Behaviour-preserving: the functions move verbatim. Full fast lane (10714
passed), ruff, mypy (0 errors), and AI-readiness (0 errors, no new import
cycle) all green.
